chore: apply code review
