+++ examples/example01-basic.html
@@ -13,7 +13,7 @@
     
     <script type="text/javascript" src="../dist/toastui-select-box.js"></script>
     <script class="code-js">
-      var selectBox = new tui.SelectBox('#select-box', {
+      const selectBox = new tui.SelectBox('#select-box', {
         placeholder: 'Please select an option.',
         data: [
           {
