+++ examples/example03-custom-events.html
@@ -22,8 +22,8 @@
     
     <script type="text/javascript" src="../dist/toastui-select-box.js"></script>
     <script class="code-js">
-      var resultDiv = document.querySelector('#result');
-      var selectBox = new tui.SelectBox('#select-box', {
+      const resultDiv = document.querySelector('#result');
+      const selectBox = new tui.SelectBox('#select-box', {
         data: [
           {
             text: 'Fruits',
@@ -47,15 +47,9 @@
         resultDiv.innerHTML += type + ': ' + text + '<br>'
       }
 
-      selectBox.on('change', function(ev) {
-        print(ev.type, ev.prev.label + ' -> ' + ev.curr.label);
-      });
-      selectBox.on('open', function(ev) {
-        print(ev.type, 'select box');
-      });
-      selectBox.on('close', function(ev) {
-        print(ev.type, 'select box');
-      });
+      selectBox.on('change', ev => print(ev.type, ev.prev.label + ' -> ' + ev.curr.label));
+      selectBox.on('open', ev => print(ev.type, 'select box'));
+      selectBox.on('close', ev => print(ev.type, 'select box'));
     </script>
   </body>
 </html>
